requirements.txt
kafka-python-ng>=2.2.3
python-dotenv==1.0.1
vaderSentiment==3.3.2
six==1.16.0
faker>=23.2.1
boto3>=1.37.11
pyarrow==15.0.0
pandas==2.2.0
psycopg2-binary==2.9.9
streamlit==1.31